Major contexts contstruction refactor into internal Window
