+++ src/index.js
@@ -153,185 +153,6 @@ const args = (() => {
   }
 })();
 
-nativeBindings.nativeGl.onconstruct = (gl, canvas) => {
-  const canvasWidth = canvas.width || innerWidth;
-  const canvasHeight = canvas.height || innerHeight;
-
-  const windowSpec = (() => {
-    try {
-      const visible = !args.image && canvas.ownerDocument.documentElement.contains(canvas);
-      const {hidden} = canvas.ownerDocument;
-      const firstWindowHandle = contexts.length > 0 ? contexts[0].getWindowHandle() : null;
-      return nativeWindow.create3d(canvasWidth, canvasHeight, visible && !hidden, firstWindowHandle, gl);
-    } catch (err) {
-      console.warn(err.message);
-      return null;
-    }
-  })();
-
-  if (windowSpec) {
-    const [windowHandle, sharedFramebuffer, sharedColorTexture, sharedDepthStencilTexture, sharedMsFramebuffer, sharedMsColorTexture, sharedMsDepthStencilTexture, vao] = windowSpec;
-
-    gl.setWindowHandle(windowHandle);
-    gl.setDefaultVao(vao);
-
-    gl.canvas = canvas;
-
-    const document = canvas.ownerDocument;
-    const window = document.defaultView;
-
-    const nativeWindowSize = nativeWindow.getFramebufferSize(windowHandle);
-    const nativeWindowHeight = nativeWindowSize.height;
-    const nativeWindowWidth = nativeWindowSize.width;
-
-    // Calculate devicePixelRatio.
-    window.devicePixelRatio = nativeWindowWidth / canvasWidth;
-
-    // Tell DOM how large the window is.
-    window.innerHeight = nativeWindowHeight / window.devicePixelRatio;
-    window.innerWidth = nativeWindowWidth / window.devicePixelRatio;
-
-    const title = `Exokit ${version}`;
-    nativeWindow.setWindowTitle(windowHandle, title);
-
-    const {hidden} = document;
-    if (hidden) {
-      const [fbo, tex, depthTex, msFbo, msTex, msDepthTex] = nativeWindow.createRenderTarget(gl, canvasWidth, canvasHeight, sharedColorTexture, sharedDepthStencilTexture, sharedMsColorTexture, sharedMsDepthStencilTexture);
-
-      gl.setDefaultFramebuffer(msFbo);
-
-      gl.resize = (width, height) => {
-        nativeWindow.setCurrentWindowContext(windowHandle);
-        nativeWindow.resizeRenderTarget(gl, width, height, fbo, tex, depthTex, msFbo, msTex, msDepthTex);
-      };
-
-      // TODO: handle multiple child canvases
-      document.framebuffer = {
-        canvas,
-        msTex,
-        msDepthTex,
-        tex,
-        depthTex,
-      };
-    } else {
-      gl.resize = (width, height) => {
-        nativeWindow.setCurrentWindowContext(windowHandle);
-        nativeWindow.resizeRenderTarget(gl, width, height, sharedFramebuffer, sharedColorTexture, sharedDepthStencilTexture, sharedMsFramebuffer, sharedMsColorTexture, sharedMsDepthStencilTexture);
-      };
-    }
-    Object.defineProperty(gl, 'drawingBufferWidth', {
-      get() {
-        return canvas.width;
-      },
-    });
-    Object.defineProperty(gl, 'drawingBufferHeight', {
-      get() {
-        return canvas.height;
-      },
-    });
-
-    const ondomchange = () => {
-      process.nextTick(() => { // show/hide synchronously emits events
-        if (!hidden) {
-          const domVisible = canvas.ownerDocument.documentElement.contains(canvas);
-          const windowVisible = nativeWindow.isVisible(windowHandle);
-          if (domVisible) {
-            if (!windowVisible) {
-              nativeWindow.show(windowHandle);
-            }
-          } else {
-            if (windowVisible) {
-              nativeWindow.hide(windowHandle);
-            }
-          }
-        }
-      });
-    };
-    canvas.ownerDocument.on('domchange', ondomchange);
-
-    gl.destroy = (destroy => function() {
-      destroy.call(this);
-
-      nativeWindow.setCurrentWindowContext(windowHandle);
-
-      if (gl === vrPresentState.glContext) {
-        nativeVr.VR_Shutdown();
-
-        vrPresentState.glContext = null;
-        vrPresentState.system = null;
-        vrPresentState.compositor = null;
-      }
-      if (gl === mlPresentState.mlGlContext) {
-        mlPresentState.mlGlContext = null;
-      }
-
-      nativeWindow.destroy(windowHandle);
-      canvas._context = null;
-
-      if (hidden) {
-        document._emit('framebuffer', null);
-      }
-      canvas.ownerDocument.removeListener('domchange', ondomchange);
-
-      contexts.splice(contexts.indexOf(gl), 1);
-
-      if (!contexts.some(context => nativeWindow.isVisible(context.getWindowHandle()))) { // no more windows
-        process.exit();
-      }
-    })(gl.destroy);
-
-    contexts.push(gl);
-    fps = nativeWindow.getRefreshRate();
-
-    canvas.ownerDocument.defaultView.on('unload', () => {
-      gl.destroy();
-    });
-
-    return true;
-  } else {
-    return false;
-  }
-};
-
-nativeBindings.nativeCanvasRenderingContext2D.onconstruct = (ctx, canvas) => {
-  const canvasWidth = canvas.width || innerWidth;
-  const canvasHeight = canvas.height || innerHeight;
-
-  const windowSpec = (() => {
-    try {
-      const firstWindowHandle = contexts.length > 0 ? contexts[0].getWindowHandle() : null;
-      return nativeWindow.create2d(canvasWidth, canvasHeight, firstWindowHandle);
-    } catch (err) {
-      console.warn(err.message);
-      return null;
-    }
-  })();
-
-  if (windowSpec) {
-    const [windowHandle, tex] = windowSpec;
-
-    ctx.setWindowHandle(windowHandle);
-    ctx.setTexture(tex, canvasWidth, canvasHeight);
-    
-    ctx.canvas = canvas;
-    
-    contexts.push(ctx);
-    
-    ctx.destroy = (destroy => function() {
-      destroy.call(this);
-      
-      nativeWindow.destroy(windowHandle);
-      canvas._context = null;
-      
-      contexts.splice(contexts.indexOf(ctx), 1);
-    })(ctx.destroy);
-
-    return true;
-  } else {
-    return false;
-  }
-};
-
 const zeroMatrix = new THREE.Matrix4();
 const localFloat32Array = zeroMatrix.toArray(new Float32Array(16));
 const localFloat32Array2 = zeroMatrix.toArray(new Float32Array(16));
